renamed RF environment
